stt/cloud-function/requirements.txt
functions-framework==3.*
google-cloud-speech==2.*
google-cloud-storage==2.*
google-genai
